Fix warnings (which are now errors) in the Android build.
BUG=None
TEST=


Review URL: http://codereview.chromium.org/9288034

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@119155 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b34a31b09919c02f620efe9974ff02f919778f78

+++ net/dns/async_host_resolver_unittest.cc
@@ -38,7 +38,8 @@ void VerifyAddressList(const std::vector<const char*>& ip_addresses,
   for (std::vector<const char*>::const_iterator i = ip_addresses.begin();
        i != ip_addresses.end(); ++i, ainfo = ainfo->ai_next) {
     ASSERT_NE(static_cast<addrinfo*>(NULL), ainfo);
-    EXPECT_EQ(sizeof(struct sockaddr_in), ainfo->ai_addrlen);
+    EXPECT_EQ(sizeof(struct sockaddr_in),
+              static_cast<size_t>(ainfo->ai_addrlen));
 
     const struct sockaddr* sa = ainfo->ai_addr;
     const struct sockaddr_in* sa_in = (const struct sockaddr_in*) sa;
